+++ erigon-lib/jsonstream/stack_stream.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with Erigon. If not, see <http://www.gnu.org/licenses/>.
 
-package json
+package jsonstream
 
 import (
 	"fmt"
