+++ rpc/handler.go
@@ -33,7 +33,7 @@ import (
 
 	jsoniter "github.com/json-iterator/go"
 
-	jsonstream "github.com/erigontech/erigon-lib/json"
+	"github.com/erigontech/erigon-lib/jsonstream"
 	"github.com/erigontech/erigon-lib/log/v3"
 
 	"github.com/erigontech/erigon/rpc/rpccfg"
